viewName validation v2
